+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 #### 功能构建
 
+- 优化了 AbstractResetter 的代码结构，同时对其实现进行了修改。
+  - com.dwarfeng.fdr.impl.handler.resetter.CronResetter。
+  - com.dwarfeng.fdr.impl.handler.resetter.DubboResetter。
+  - com.dwarfeng.fdr.impl.handler.resetter.FixedDelayResetter。
+  - com.dwarfeng.fdr.impl.handler.resetter.FixedRateResetter。
+  - com.dwarfeng.fdr.impl.handler.resetter.NeverResetter。
+
 - 优化部分指令的文案。
   - com.dwarfeng.fdr.impl.service.telqos.LogicConsumerCommand。
   - com.dwarfeng.fdr.impl.service.telqos.RecordConsumerCommand。